requirements.txt
markdown
click
mkdocs-click
click-logging
pyyaml
semver
pytest
requests
pyjwt